+++ eActForm/eActForm/BusinessLayer/Appcodes/ImageAppCodeBudget.cs
@@ -9,278 +9,279 @@ using WebLibrary;
 
 namespace eActForm.BusinessLayer
 {
-    public class ImageAppCodeBudget
-    {
+    //public class ImageAppCodeBudget
+    //{
 
-        //public static List<TB_Bud_Image_Model.BudImageModel> getImageBudgetByApproveId(string budgetApproveId)
-        public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoiceByApproveId(string budgetApproveId)
-        {
-            try
-            {
-                DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetApproveInvoiceList"
-                    , new SqlParameter("@budgetApproveId", budgetApproveId)
-                    //, new SqlParameter("@activityNo", activityNo)
-                    //, new SqlParameter("@createdByUserId", createdByUserId)
-                    );
-                var lists = (from DataRow d in ds.Tables[0].Rows
-                             select new TB_Bud_Image_Model.BudImageModel()
-                             {
-                                 id = d["imageId"].ToString(),
+    //    //public static List<TB_Bud_Image_Model.BudImageModel> getImageBudgetByApproveId(string budgetApproveId)
+    //    public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoiceByApproveId(string budgetApproveId)
+    //    {
+    //        try
+    //        {
+    //            DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetApproveInvoiceList"
+    //                , new SqlParameter("@budgetApproveId", budgetApproveId)
+    //                //, new SqlParameter("@activityNo", activityNo)
+    //                //, new SqlParameter("@createdByUserId", createdByUserId)
+    //                );
+    //            var lists = (from DataRow d in ds.Tables[0].Rows
+    //                         select new TB_Bud_Image_Model.BudImageModel()
+    //                         {
+    //                             id = d["imageId"].ToString(),
 
-                                 invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
-                                 imageType = d["imageType"].ToString(),
-                                 _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
-                                 _fileName = d["_fileName"].ToString(),
-                                 extension = d["extension"].ToString(),
-                                 remark = d["remark"].ToString(),
-                                 delFlag = bool.Parse(d["delFlag"].ToString()),
-                                 createdDate = DateTime.Parse(d["createdDate"].ToString()),
-                                 createdByUserId = d["createdByUserId"].ToString(),
-                                 updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
-                                 updatedByUserId = d["updatedByUserId"].ToString(),
-                             });
-                return lists.ToList();
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError("getBudgetInvoiceByApproveId => " + ex.Message);
-                return new List<TB_Bud_Image_Model.BudImageModel>();
-            }
-        }
-
-
-        //public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoiceList(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId)
-        //{
-        //    try
-        //    {
-        //        DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_getBudgetImage"
-        //            , new SqlParameter("@imageId", imageId)
-        //            , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
-        //            , new SqlParameter("@budgetApproveId", budgetApproveId)
-        //            , new SqlParameter("@activityNo", activityNo)
-        //            , new SqlParameter("@createdByUserId", createdByUserId)
-        //            , new SqlParameter("@company", company)
-        //            , new SqlParameter("@customerId", customerId)
-        //            );
-        //        var lists = (from DataRow d in ds.Tables[0].Rows
-        //                     select new TB_Bud_Image_Model.BudImageModel()
-        //                     {
-        //                         id = d["imageId"].ToString(),
-
-        //                         count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
-        //                         count_activityNo = int.Parse(d["count_activityNo"].ToString()),
-        //                         count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
-
-        //                         invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
-        //                         imageType = d["imageType"].ToString(),
-        //                         _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
-        //                         _fileName = d["_fileName"].ToString(),
-        //                         extension = d["extension"].ToString(),
-        //                         remark = d["remark"].ToString(),
-
-        //                         companyId = d["companyId"].ToString(),
-        //                         regionId = d["regionId"].ToString(),
-        //                         customerId = d["customerId"].ToString(),
-
-        //                         company = d["company"].ToString(),
-        //                         regionName = d["regionName"].ToString(),
-        //                         customerName = d["cusNameTH"].ToString(),
-
-        //                         delFlag = bool.Parse(d["delFlag"].ToString()),
-        //                         createdDate = DateTime.Parse(d["createdDate"].ToString()),
-        //                         createdByUserId = d["createdByUserId"].ToString(),
-        //                         updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
-        //                         updatedByUserId = d["updatedByUserId"].ToString(),
-
-        //                     });
-        //        return lists.ToList();
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        ExceptionManager.WriteError("getImage budget => " + ex.Message);
-        //        return new List<TB_Bud_Image_Model.BudImageModel>();
-        //    }
-        //}
-
-        //public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoice(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId)
-        //{
-        //    try
-        //    {
-        //        DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_getBudgetImage"
-        //            , new SqlParameter("@imageId", imageId)
-        //            , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
-        //            , new SqlParameter("@budgetApproveId", budgetApproveId)
-        //            , new SqlParameter("@activityNo", activityNo)
-        //            , new SqlParameter("@createdByUserId", createdByUserId)
-        //            , new SqlParameter("@company", company)
-        //            , new SqlParameter("@customerId", customerId)
-        //            );
-        //        var lists = (from DataRow d in ds.Tables[0].Rows
-        //                     select new TB_Bud_Image_Model.BudImageModel()
-        //                     {
-        //                         id = d["imageId"].ToString(),
-
-        //                         count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
-        //                         count_activityNo = int.Parse(d["count_activityNo"].ToString()),
-        //                         count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
-
-        //                         invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
-        //                         imageType = d["imageType"].ToString(),
-        //                         _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
-        //                         _fileName = d["_fileName"].ToString(),
-        //                         extension = d["extension"].ToString(),
-        //                         remark = d["remark"].ToString(),
-
-        //                         companyId = d["companyId"].ToString(),
-        //                         regionId = d["regionId"].ToString(),
-        //                         customerId = d["customerId"].ToString(),
-
-        //                         company = d["company"].ToString(),
-        //                         regionName = d["regionName"].ToString(),
-        //                         customerName = d["cusNameTH"].ToString(),
-
-        //                         delFlag = bool.Parse(d["delFlag"].ToString()),
-        //                         createdDate = DateTime.Parse(d["createdDate"].ToString()),
-        //                         createdByUserId = d["createdByUserId"].ToString(),
-        //                         updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
-        //                         updatedByUserId = d["updatedByUserId"].ToString(),
-
-        //                     });
-        //        return lists.ToList();
-        //    }
-        //    catch (Exception ex)
-        //    {
-        //        ExceptionManager.WriteError("getImage budget => " + ex.Message);
-        //        return new List<TB_Bud_Image_Model.BudImageModel>();
-        //    }
-        //}
+    //                             invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
+    //                             imageType = d["imageType"].ToString(),
+    //                             _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
+    //                             _fileName = d["_fileName"].ToString(),
+    //                             extension = d["extension"].ToString(),
+    //                             remark = d["remark"].ToString(),
+    //                             delFlag = bool.Parse(d["delFlag"].ToString()),
+    //                             createdDate = DateTime.Parse(d["createdDate"].ToString()),
+    //                             createdByUserId = d["createdByUserId"].ToString(),
+    //                             updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
+    //                             updatedByUserId = d["updatedByUserId"].ToString(),
+    //                         });
+    //            return lists.ToList();
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError("getBudgetInvoiceByApproveId => " + ex.Message);
+    //            return new List<TB_Bud_Image_Model.BudImageModel>();
+    //        }
+    //    }
 
 
-        public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoice(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId, string beginDateyyyymmdd, string endDateyyyymmdd)
-        {
-            try
-            {
-                DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceList"
-                    , new SqlParameter("@imageId", imageId)
-                    , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
-                    , new SqlParameter("@budgetApproveId", budgetApproveId)
-                    , new SqlParameter("@activityNo", activityNo)
-                    , new SqlParameter("@createdByUserId", createdByUserId)
-                    , new SqlParameter("@company", company)
-                    , new SqlParameter("@customerId", customerId)
-                    //, new SqlParameter("@beginDateyyyymmdd", null)
-                    //, new SqlParameter("@endDateyyyymmdd", null)
-                    , new SqlParameter("@beginDateyyyymmdd", beginDateyyyymmdd)
-                    , new SqlParameter("@endDateyyyymmdd", endDateyyyymmdd)
-                    );
-                var lists = (from DataRow d in ds.Tables[0].Rows
-                             select new TB_Bud_Image_Model.BudImageModel()
-                             {
-                                 id = d["imageId"].ToString(),
+    //    //public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoiceList(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId)
+    //    //{
+    //    //    try
+    //    //    {
+    //    //        DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_getBudgetImage"
+    //    //            , new SqlParameter("@imageId", imageId)
+    //    //            , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
+    //    //            , new SqlParameter("@budgetApproveId", budgetApproveId)
+    //    //            , new SqlParameter("@activityNo", activityNo)
+    //    //            , new SqlParameter("@createdByUserId", createdByUserId)
+    //    //            , new SqlParameter("@company", company)
+    //    //            , new SqlParameter("@customerId", customerId)
+    //    //            );
+    //    //        var lists = (from DataRow d in ds.Tables[0].Rows
+    //    //                     select new TB_Bud_Image_Model.BudImageModel()
+    //    //                     {
+    //    //                         id = d["imageId"].ToString(),
 
-                                 count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
-                                 count_activityNo = int.Parse(d["count_activityNo"].ToString()),
-                                 count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
+    //    //                         count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
+    //    //                         count_activityNo = int.Parse(d["count_activityNo"].ToString()),
+    //    //                         count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
 
-                                 invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
-                                 imageType = d["imageType"].ToString(),
-                                 _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
-                                 _fileName = d["_fileName"].ToString(),
-                                 extension = d["extension"].ToString(),
-                                 remark = d["remark"].ToString(),
+    //    //                         invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
+    //    //                         imageType = d["imageType"].ToString(),
+    //    //                         _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
+    //    //                         _fileName = d["_fileName"].ToString(),
+    //    //                         extension = d["extension"].ToString(),
+    //    //                         remark = d["remark"].ToString(),
 
-                                 companyId = d["companyId"].ToString(),
-                                 regionId = d["regionId"].ToString(),
-                                 customerId = d["customerId"].ToString(),
+    //    //                         companyId = d["companyId"].ToString(),
+    //    //                         regionId = d["regionId"].ToString(),
+    //    //                         customerId = d["customerId"].ToString(),
 
-                                 company = d["company"].ToString(),
-                                 regionName = d["regionName"].ToString(),
-                                 customerName = d["cusNameTH"].ToString(),
+    //    //                         company = d["company"].ToString(),
+    //    //                         regionName = d["regionName"].ToString(),
+    //    //                         customerName = d["cusNameTH"].ToString(),
 
-                                 delFlag = bool.Parse(d["delFlag"].ToString()),
-                                 createdDate = DateTime.Parse(d["createdDate"].ToString()),
-                                 createdByUserId = d["createdByUserId"].ToString(),
-                                 updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
-                                 updatedByUserId = d["updatedByUserId"].ToString(),
+    //    //                         delFlag = bool.Parse(d["delFlag"].ToString()),
+    //    //                         createdDate = DateTime.Parse(d["createdDate"].ToString()),
+    //    //                         createdByUserId = d["createdByUserId"].ToString(),
+    //    //                         updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
+    //    //                         updatedByUserId = d["updatedByUserId"].ToString(),
 
-                             });
-                return lists.ToList();
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError("getBudgetInvoice => " + ex.Message);
-                return new List<TB_Bud_Image_Model.BudImageModel>();
-            }
-        }
+    //    //                     });
+    //    //        return lists.ToList();
+    //    //    }
+    //    //    catch (Exception ex)
+    //    //    {
+    //    //        ExceptionManager.WriteError("getImage budget => " + ex.Message);
+    //    //        return new List<TB_Bud_Image_Model.BudImageModel>();
+    //    //    }
+    //    //}
 
-        public static int insertBudgetInvoice(TB_Bud_Image_Model.BudImageModel model)
-        {
-            int result = 0;
-            try
-            {
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceInsert"
-                    , new SqlParameter[] {new SqlParameter("@imageType",model.imageType)
-                    ,new SqlParameter("@image",model._image)
-                    ,new SqlParameter("@fileName",model._fileName)
-                    ,new SqlParameter("@extension",model.extension)
-                    ,new SqlParameter("@remark",model.remark)
-                    ,new SqlParameter("@company",model.company)
-                    ,new SqlParameter("@delFlag",model.delFlag)
-                    ,new SqlParameter("@createdDate",model.createdDate)
-                    ,new SqlParameter("@createdByUserId",model.createdByUserId)
-                    ,new SqlParameter("@updatedDate",model.updatedDate)
-                    ,new SqlParameter("@updatedByUserId",model.updatedByUserId)
+    //    //public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoice(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId)
+    //    //{
+    //    //    try
+    //    //    {
+    //    //        DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_getBudgetImage"
+    //    //            , new SqlParameter("@imageId", imageId)
+    //    //            , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
+    //    //            , new SqlParameter("@budgetApproveId", budgetApproveId)
+    //    //            , new SqlParameter("@activityNo", activityNo)
+    //    //            , new SqlParameter("@createdByUserId", createdByUserId)
+    //    //            , new SqlParameter("@company", company)
+    //    //            , new SqlParameter("@customerId", customerId)
+    //    //            );
+    //    //        var lists = (from DataRow d in ds.Tables[0].Rows
+    //    //                     select new TB_Bud_Image_Model.BudImageModel()
+    //    //                     {
+    //    //                         id = d["imageId"].ToString(),
 
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> insertBudgetInvoice");
-            }
+    //    //                         count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
+    //    //                         count_activityNo = int.Parse(d["count_activityNo"].ToString()),
+    //    //                         count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
 
-            return result;
-        }
+    //    //                         invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
+    //    //                         imageType = d["imageType"].ToString(),
+    //    //                         _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
+    //    //                         _fileName = d["_fileName"].ToString(),
+    //    //                         extension = d["extension"].ToString(),
+    //    //                         remark = d["remark"].ToString(),
 
-        public static int updateImageBudget(TB_Bud_Image_Model.BudImageModel model)
-        {
-            int result = 0;
-            try
-            {
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceUpdate"
-                    , new SqlParameter[] {new SqlParameter("@id",model.id)
-                    ,new SqlParameter("@invoiceNo",model.invoiceNo)
-                    ,new SqlParameter("@customerId",model.customerId)
-                    ,new SqlParameter("@remark",model.remark)
-                    ,new SqlParameter("@updatedByUserId",model.updatedByUserId)
+    //    //                         companyId = d["companyId"].ToString(),
+    //    //                         regionId = d["regionId"].ToString(),
+    //    //                         customerId = d["customerId"].ToString(),
 
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> updateImageBudget");
-            }
+    //    //                         company = d["company"].ToString(),
+    //    //                         regionName = d["regionName"].ToString(),
+    //    //                         customerName = d["cusNameTH"].ToString(),
 
-            return result;
-        }
+    //    //                         delFlag = bool.Parse(d["delFlag"].ToString()),
+    //    //                         createdDate = DateTime.Parse(d["createdDate"].ToString()),
+    //    //                         createdByUserId = d["createdByUserId"].ToString(),
+    //    //                         updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
+    //    //                         updatedByUserId = d["updatedByUserId"].ToString(),
 
-        public static int deleteImageBudgetById(string fileId, string empId)
-        {
+    //    //                     });
+    //    //        return lists.ToList();
+    //    //    }
+    //    //    catch (Exception ex)
+    //    //    {
+    //    //        ExceptionManager.WriteError("getImage budget => " + ex.Message);
+    //    //        return new List<TB_Bud_Image_Model.BudImageModel>();
+    //    //    }
+    //    //}
 
-            int result = 0;
-            try
-            {
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceDelete"
-                    , new SqlParameter[] {new SqlParameter("@Id",fileId)
-                    ,new SqlParameter("@updatedByUserId",empId)
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> deleteImageBudgetById");
-            }
 
-            return result;
-        }
+    //    public static List<TB_Bud_Image_Model.BudImageModel> getBudgetInvoice(string imageId, string imageInvoiceNo, string budgetApproveId, string activityNo, string createdByUserId, string company, string customerId, string beginDateyyyymmdd, string endDateyyyymmdd)
+    //    {
+    //        try
+    //        {
+    //            DataSet ds = SqlHelper.ExecuteDataset(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceList"
+    //                , new SqlParameter("@imageId", imageId)
+    //                , new SqlParameter("@imageInvoiceNo", imageInvoiceNo)
+    //                , new SqlParameter("@budgetApproveId", budgetApproveId)
+    //                , new SqlParameter("@activityNo", activityNo)
+    //                , new SqlParameter("@createdByUserId", createdByUserId)
+    //                , new SqlParameter("@company", company)
+    //                , new SqlParameter("@customerId", customerId)
+    //                //, new SqlParameter("@beginDateyyyymmdd", null)
+    //                //, new SqlParameter("@endDateyyyymmdd", null)
+    //                , new SqlParameter("@beginDateyyyymmdd", beginDateyyyymmdd)
+    //                , new SqlParameter("@endDateyyyymmdd", endDateyyyymmdd)
+    //                );
+    //            var lists = (from DataRow d in ds.Tables[0].Rows
+    //                         select new TB_Bud_Image_Model.BudImageModel()
+    //                         {
+    //                             id = d["imageId"].ToString(),
 
-    }
+    //                             count_budgetActivityId = int.Parse(d["count_budgetActivityId"].ToString()),
+    //                             count_activityNo = int.Parse(d["count_activityNo"].ToString()),
+    //                             count_budgetApproveId = int.Parse(d["count_budgetApproveId"].ToString()),
+
+    //                             invoiceNo = (d["invoiceNo"].ToString() == null || d["invoiceNo"] is DBNull) ? "" : d["invoiceNo"].ToString(),
+    //                             imageType = d["imageType"].ToString(),
+    //                             _image = (d["_image"] == null || d["_image"] is DBNull) ? new byte[0] : (byte[])d["_image"],
+    //                             _fileName = d["_fileName"].ToString(),
+    //                             extension = d["extension"].ToString(),
+    //                             remark = d["remark"].ToString(),
+
+    //                             companyId = d["companyId"].ToString(),
+    //                             regionId = d["regionId"].ToString(),
+    //                             customerId = d["customerId"].ToString(),
+
+    //                             company = d["company"].ToString(),
+    //                             regionName = d["regionName"].ToString(),
+    //                             customerName = d["cusNameTH"].ToString(),
+
+    //                             delFlag = bool.Parse(d["delFlag"].ToString()),
+    //                             createdDate = DateTime.Parse(d["createdDate"].ToString()),
+    //                             createdByUserId = d["createdByUserId"].ToString(),
+    //                             updatedDate = DateTime.Parse(d["updatedDate"].ToString()),
+    //                             updatedByUserId = d["updatedByUserId"].ToString(),
+
+    //                         });
+    //            return lists.ToList();
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError("getBudgetInvoice => " + ex.Message);
+    //            return new List<TB_Bud_Image_Model.BudImageModel>();
+    //        }
+    //    }
+
+    //    public static int insertBudgetInvoice(TB_Bud_Image_Model.BudImageModel model)
+    //    {
+    //        int result = 0;
+    //        try
+    //        {
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceInsert"
+    //                , new SqlParameter[] {new SqlParameter("@imageType",model.imageType)
+    //                ,new SqlParameter("@image",model._image)
+    //                ,new SqlParameter("@fileName",model._fileName)
+    //                ,new SqlParameter("@extension",model.extension)
+    //                ,new SqlParameter("@remark",model.remark)
+    //                ,new SqlParameter("@company",model.company)
+    //                ,new SqlParameter("@delFlag",model.delFlag)
+    //                ,new SqlParameter("@createdDate",model.createdDate)
+    //                ,new SqlParameter("@createdByUserId",model.createdByUserId)
+    //                ,new SqlParameter("@updatedDate",model.updatedDate)
+    //                ,new SqlParameter("@updatedByUserId",model.updatedByUserId)
+
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> insertBudgetInvoice");
+    //        }
+
+    //        return result;
+    //    }
+
+    //    public static int updateImageBudget(TB_Bud_Image_Model.BudImageModel model)
+    //    {
+    //        int result = 0;
+    //        try
+    //        {
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceUpdate"
+    //                , new SqlParameter[] {new SqlParameter("@id",model.id)
+    //                ,new SqlParameter("@invoiceNo",model.invoiceNo)
+    //                ,new SqlParameter("@customerId",model.customerId)
+    //                ,new SqlParameter("@remark",model.remark)
+    //                ,new SqlParameter("@updatedByUserId",model.updatedByUserId)
+
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> updateImageBudget");
+    //        }
+
+    //        return result;
+    //    }
+
+    //    public static int deleteImageBudgetById(string fileId, string empId)
+    //    {
+
+    //        int result = 0;
+    //        try
+    //        {
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetInvoiceDelete"
+    //                , new SqlParameter[] {new SqlParameter("@Id",fileId)
+    //                ,new SqlParameter("@updatedByUserId",empId)
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> deleteImageBudgetById");
+    //        }
+
+    //        return result;
+    //    }
+
+    //}
+
 }
