+++ eActForm/eActForm/BusinessLayer/CommandHandler/BudgetFormCommandHandler.cs
@@ -8,125 +8,126 @@ using WebLibrary;
 
 namespace eActForm.BusinessLayer
 {
-    public class BudgetFormCommandHandler
-    {
-        //update Invoice Product
-        public static int updateInvoiceProduct(Budget_Activity_Model.Budget_Activity_Invoice_Att model)
-        {
-            int result = 0;
-            model.dateInvoiceAction = DateTime.ParseExact(model.invoiceActionDate, "dd-MM-yyyy", CultureInfo.InvariantCulture);
+    //public class BudgetFormCommandHandler
+    //{
+    //    //update Invoice Product
+    //    public static int updateInvoiceProduct(Budget_Activity_Model.Budget_Activity_Invoice_Att model)
+    //    {
+    //        int result = 0;
+    //        model.dateInvoiceAction = DateTime.ParseExact(model.invoiceActionDate, "dd-MM-yyyy", CultureInfo.InvariantCulture);
 
-            try
-            {
+    //        try
+    //        {
 
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceUpdate"
-                    , new SqlParameter[] {new SqlParameter("@id", model.invoiceId)
-                    ,new SqlParameter("@activityId",model.activityId)
-                    ,new SqlParameter("@activityNo",model.activityNo)
-                    ,new SqlParameter("@productId",model.productId)
-                    ,new SqlParameter("@activityOfEstimateId",model.activityOfEstimateId)
-                    ,new SqlParameter("@paymentNo",model.paymentNo)
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceUpdate"
+    //                , new SqlParameter[] {new SqlParameter("@id", model.invoiceId)
+    //                ,new SqlParameter("@activityId",model.activityId)
+    //                ,new SqlParameter("@activityNo",model.activityNo)
+    //                ,new SqlParameter("@productId",model.productId)
+    //                ,new SqlParameter("@activityOfEstimateId",model.activityOfEstimateId)
+    //                ,new SqlParameter("@paymentNo",model.paymentNo)
 
-                    ,new SqlParameter("@budgetImageId",model.budgetImageId)
+    //                ,new SqlParameter("@budgetImageId",model.budgetImageId)
 
-                    ,new SqlParameter("@invoiceBudgetStatusId",model.invoiceBudgetStatusId)
-                    ,new SqlParameter("@invoiceNo",model.invoiceNo)
-                    ,new SqlParameter("@invoiceTotalBath",model.invoiceTotalBath)
-                    ,new SqlParameter("@actionDate",model.dateInvoiceAction) //invoiceActionDate
-                    ,new SqlParameter("@invoiceRemark",model.invoiceRemark)
+    //                ,new SqlParameter("@invoiceBudgetStatusId",model.invoiceBudgetStatusId)
+    //                ,new SqlParameter("@invoiceNo",model.invoiceNo)
+    //                ,new SqlParameter("@invoiceTotalBath",model.invoiceTotalBath)
+    //                ,new SqlParameter("@actionDate",model.dateInvoiceAction) //invoiceActionDate
+    //                ,new SqlParameter("@invoiceRemark",model.invoiceRemark)
 
-                    ,new SqlParameter("@createdByUserId",UtilsAppCode.Session.User.empId)
-                    ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceUpdate");
-            }
+    //                ,new SqlParameter("@createdByUserId",UtilsAppCode.Session.User.empId)
+    //                ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceUpdate");
+    //        }
 
-            return result;
-        }
+    //        return result;
+    //    }
 
-        public static int insertInvoiceProduct(Budget_Activity_Model.Budget_Activity_Invoice_Att model)
-        {
+    //    public static int insertInvoiceProduct(Budget_Activity_Model.Budget_Activity_Invoice_Att model)
+    //    {
 
-            int result = 0;
-            model.dateInvoiceAction = DateTime.ParseExact(model.invoiceActionDate, "dd-MM-yyyy", CultureInfo.InvariantCulture);
+    //        int result = 0;
+    //        model.dateInvoiceAction = DateTime.ParseExact(model.invoiceActionDate, "dd-MM-yyyy", CultureInfo.InvariantCulture);
 
-            try
-            {
+    //        try
+    //        {
 
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceInsert"
-                    , new SqlParameter[] {new SqlParameter("@id", Guid.NewGuid().ToString())
-                    ,new SqlParameter("@activityId",model.activityId)
-                    ,new SqlParameter("@activityNo",model.activityNo)
-                    ,new SqlParameter("@productId",model.productId)
-                    ,new SqlParameter("@activityOfEstimateId",model.activityOfEstimateId)
-                    ,new SqlParameter("@paymentNo",model.paymentNo)
-                    ,new SqlParameter("@budgetImageId",model.budgetImageId)
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceInsert"
+    //                , new SqlParameter[] {new SqlParameter("@id", Guid.NewGuid().ToString())
+    //                ,new SqlParameter("@activityId",model.activityId)
+    //                ,new SqlParameter("@activityNo",model.activityNo)
+    //                ,new SqlParameter("@productId",model.productId)
+    //                ,new SqlParameter("@activityOfEstimateId",model.activityOfEstimateId)
+    //                ,new SqlParameter("@paymentNo",model.paymentNo)
+    //                ,new SqlParameter("@budgetImageId",model.budgetImageId)
 
-                    ,new SqlParameter("@invoiceBudgetStatusId",model.invoiceBudgetStatusId)
-                    ,new SqlParameter("@invoiceNo",model.invoiceNo)
-                    ,new SqlParameter("@invoiceTotalBath",model.invoiceTotalBath)
-					//,new SqlParameter("@actionDate",model.invoiceActionDate)
-					,new SqlParameter("@actionDate",model.dateInvoiceAction)
-                    ,new SqlParameter("@invoiceRemark",model.invoiceRemark)
-                    ,new SqlParameter("@createdByUserId",UtilsAppCode.Session.User.empId)
-                    ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceInsert");
-            }
+    //                ,new SqlParameter("@invoiceBudgetStatusId",model.invoiceBudgetStatusId)
+    //                ,new SqlParameter("@invoiceNo",model.invoiceNo)
+    //                ,new SqlParameter("@invoiceTotalBath",model.invoiceTotalBath)
+				//	//,new SqlParameter("@actionDate",model.invoiceActionDate)
+				//	,new SqlParameter("@actionDate",model.dateInvoiceAction)
+    //                ,new SqlParameter("@invoiceRemark",model.invoiceRemark)
+    //                ,new SqlParameter("@createdByUserId",UtilsAppCode.Session.User.empId)
+    //                ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceInsert");
+    //        }
 
-            return result;
-        }
+    //        return result;
+    //    }
 
-        public static int deleteInvoiceProduct(string activityId, string estimateId, string invoiceId, string delType)
-        //public static int deleteInvoiceProduct(string activityId, string estimateId, string invoiceNo)
-        {
+    //    public static int deleteInvoiceProduct(string activityId, string estimateId, string invoiceId, string delType)
+    //    //public static int deleteInvoiceProduct(string activityId, string estimateId, string invoiceNo)
+    //    {
 
-            int result = 0;
+    //        int result = 0;
 
-            try
-            {
+    //        try
+    //        {
 
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceDelete"
-                    , new SqlParameter[] {new SqlParameter("@activityId",activityId)
-                    ,new SqlParameter("@activityOfEstimateId",estimateId)
-                    ,new SqlParameter("@invoiceId",invoiceId)
-                    ,new SqlParameter("@delType",delType)
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceDelete");
-            }
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetActivityInvoiceDelete"
+    //                , new SqlParameter[] {new SqlParameter("@activityId",activityId)
+    //                ,new SqlParameter("@activityOfEstimateId",estimateId)
+    //                ,new SqlParameter("@invoiceId",invoiceId)
+    //                ,new SqlParameter("@delType",delType)
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> usp_mtm_BudgetActivityInvoiceDelete");
+    //        }
 
-            return result;
-        }
+    //        return result;
+    //    }
 
-        public static int deleteBudgetApproveByActNo(string activityNo)
-        {
+    //    public static int deleteBudgetApproveByActNo(string activityNo)
+    //    {
 
-            int result = 0;
+    //        int result = 0;
 
-            try
-            {
+    //        try
+    //        {
 
-                result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetApproveDelete"
-                    , new SqlParameter[] {new SqlParameter("@activityNo",activityNo)
-                    ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
-                    });
-            }
-            catch (Exception ex)
-            {
-                ExceptionManager.WriteError(ex.Message + ">> deleteBudgetApproveByActNo");
-            }
+    //            result = SqlHelper.ExecuteNonQuery(AppCode.StrCon, CommandType.StoredProcedure, "usp_mtm_BudgetApproveDelete"
+    //                , new SqlParameter[] {new SqlParameter("@activityNo",activityNo)
+    //                ,new SqlParameter("@updatedByUserId",UtilsAppCode.Session.User.empId)
+    //                });
+    //        }
+    //        catch (Exception ex)
+    //        {
+    //            ExceptionManager.WriteError(ex.Message + ">> deleteBudgetApproveByActNo");
+    //        }
 
-            return result;
-        }
+    //        return result;
+    //    }
 
-    }
+    //}
+
 }
