+++ eActForm/eActForm/Views/Budget/manageInvoiceEdit.cshtml
@@ -1,4 +1,4 @@
-﻿@model eActForm.Models.TB_Bud_Image_Model.BudImageModel
+﻿@model eActForm.Models.TB_Bud_Invoice_Document_Model.BudgetInvoiceModel
 @{
     //update 21-04-2020
     Layout = null;
