+++ eActForm/eActForm/Views/Shared/_searchBudgetInvoice.cshtml
@@ -1,7 +1,7 @@
 ﻿@*@model eActForm.Models.Budget_Activity_Model*@
 @using eActForm.BusinessLayer;
 @using System.Configuration;
-@model eActForm.Models.TB_Bud_Image_Model
+@model eActForm.Models.TB_Bud_Invoice_Document_Model
 
 <!--<div class="form-group">
     <div class="col-md-6 col-xs-12">
